requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn==0.27.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Email (Gmail API)
google-auth==2.27.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.119.0
email-validator==2.1.0

# Attachment Parsing
pymupdf==1.23.8
python-docx==1.1.0

# AI Integration
httpx==0.26.0
openpyxl==3.1.5

# Security
bcrypt==4.1.2
itsdangerous==2.1.2

# Environment Variables
python-dotenv==1.0.0

# Jinja2 for templates
jinja2==3.1.2

# Password hashing
passlib==1.7.4

# CORS middleware (already in fastapi)
# email-validator already included
python-multipart==0.0.27

# Background Task Scheduler
apscheduler==3.10.4

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Production utilities
gunicorn==21.2.0
python-json-logger==2.0.7
sentry-sdk==1.40.0
